requirements.txt
streamlit
supabase
google-generativeai>=0.5.0
pytz
streamlit-cookies-manager